refactor(math): umt_gcd関数の引数型をi64からi32に変更

+++ package/umt_rust/src/math/gcd.rs
@@ -1,4 +1,4 @@
-pub fn umt_gcd(mut x: i64, mut y: i64) -> i64 {
+pub fn umt_gcd(mut x: i32, mut y: i32) -> i32 {
     while y != 0 {
         let temp = y;
         y = x % y;
